Update the user credential cache documentation
The release notes now mention that the caching is done by the
authenticators, not by the core. It also mentions that only MySQLAuth
implements this functionality.

Added notes to binlogrouter documentation about the changes to the loading
and storing of users.

Cleaned up MySQL-Authenticator document to make it a bit easier to read.