Fix memory leaks in maxinfo (v1.4.4)
MXS-1009. This commit adds a gwbuf_free after maxinfo_execute() to
free a buffer with an sql-query after it has been processed. Also,
the parse tree in maxinfo_execute_query() is now freed. The tree_free-
function was renamed to maxinfo_tree_free, since it is now globally
available.

Finally, addrlen is initialized now, it doesn't seem to matter but
valgrind no longer complains about it.